T_invest_api/requirements.txt
fastapi[standard]
python-dotenv
aiohttp
tinkoff-investments